requirements.txt
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.2
pytz==2024.1
